Share cloud-provider configuration through BOSH links
This change updates the cloud-provider to be consumed as a BOSH link.
Currently jobs will template these properties into their own templates,
but the template files are shared using symlinks.

This change allows us the ability to complete the conversion of the
kube-apiserver and kube-controller-manager to BPM as the jobs no longer
depend on templates from the `cloud-provider`.

[#153690804]

Signed-off-by: James Myers <jfmyers9@gmail.com>